Move deploy_token_contract into ContractDeployer
This is for calming down CodeClimate
https://github.com/raiden-network/raiden-contracts/issues/714
